[SYCL] Check cache directory settings
Add tests to check that:
 - cache is disabled if no environment variables defining cache root
directory are set;
 - cache can be enabled if at least one evironment variable is set.
